Introduces MavenRepository interface and default implementation
This is a breaking change, so the major version number have been bumped as a result of this change.

This is being done to make it easier to test certain scenarios where the target repository and Maven central may return different results. Tests for that scenario will come in further commits.